fix: US-001 - Split search.rs into search/pipeline.rs module
Extract pipeline helpers (resolve_sparse_input, parse_fusion_strategy,
execute_dense_search, execute_search_request, finish_search,
finish_search_ids, validate_query_dimension, actionable_search_error,
dimension_mismatch_error) into handlers/search/pipeline.rs.

Convert handlers/search.rs into handlers/search/mod.rs retaining only
the public handler functions. mod.rs: 447 NLOC, pipeline.rs: 272 NLOC.
